# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water App",
    "license": {
      "name": "Apache 2.0",
      "url": "<http://www.apache.org/licenses/LICENSE-2.0.html>"
    },
    "description": "This a documentation of my Water App\n"
  },
  "servers": [
    {
      "url": "https://water-app-b.onrender.com"
    }
  ],
  "tags": [
    {
      "name": "Water",
      "description": "Operation with used water"
    },
    {
      "name": "Auth",
      "description": "Auth operations"
    }
  ],
  "paths": {
    "/water": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get Water Info",
        "operationId": "getAllWater",
        "description": "Get water info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "day",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "string",
              "example": 2024
            },
            "description": "All fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found used water!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found used water!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "items",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasPreviousPage",
                        "hasNextPage"
                      ],
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/water"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 30
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 30
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": ["Water"],
        "summary": "Create used water",
        "operationId": "addSpent",
        "description": "Add a used water with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ml", "time", "day", "month"],
                "properties": {
                  "ml": {
                    "description": "Ml drunk",
                    "type": "string",
                    "example": "350"
                  },
                  "time": {
                    "description": "Time when the water was drunk",
                    "type": "string",
                    "example": "10:00"
                  },
                  "day": {
                    "description": "Day when the water was drunk",
                    "type": "string",
                    "example": "5"
                  },
                  "month": {
                    "description": "Month when the water was drunk",
                    "type": "string",
                    "example": "5"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a used water",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully added info about used water!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/water/{id}": {
      "get": {
        "tags": ["Water"],
        "summary": "Get water by id",
        "operationId": "getSpentById",
        "description": "Get water by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get water by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found information about used with id 65ca67e7ae7f10c88b598384!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": ["Water"],
        "summary": "Update water",
        "operationId": "upsertSpent",
        "description": "Update a used water by ID with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the water used to update"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ml": {
                    "description": "Ml drunk",
                    "type": "string",
                    "example": "350"
                  },
                  "time": {
                    "description": "Time when the water was drunk",
                    "type": "string",
                    "example": "10:00"
                  },
                  "day": {
                    "description": "Day when the water was drunk",
                    "type": "string",
                    "example": "5"
                  },
                  "month": {
                    "description": "Month when the water was drunk",
                    "type": "string",
                    "example": "5"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated information about used water!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated information about used water!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/water"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete water",
        "operationId": "deleteSpent",
        "description": "Delete a water by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the water to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted information about used water!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted information about used water!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "water": {
        "type": "object",
        "required": ["_id", "name", "phoneNumber", "email", "contactType"],
        "properties": {
          "_id": {
            "description": "Contact's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "ml": {
            "description": "Ml drunk",
            "type": "string",
            "example": "350"
          },
          "day": {
            "description": "Day when the water was drunk",
            "type": "string",
            "example": "5"
          },
          "monthNumber": {
            "description": "Month when the water was drunk",
            "type": "string",
            "example": "5"
          },
          "year": {
            "description": "Year when the water was drunk",
            "type": "string",
            "example": "2024"
          },
          "time": {
            "description": "Time when the water was drunk",
            "type": "string",
            "example": "10:00"
          }
        }
      }
    },
    "responses": {
      "404": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}